relax scalacOptions on scalafix invocations even without scalafixEnable
