+++ src/sbt-test/sbt-scalafix/relax-scalacOptions/src/main/scala/UnusedImportAndProcedureSyntax.scala
