+++ src/sbt-test/sbt-scalafix/scalafixEnable/build.sbt
@@ -21,31 +21,7 @@ lazy val scala211 = project.settings(
 
 // 2.12.x is supported
 lazy val scala212 = project.settings(
-  scalaVersion := V.scala212,
-  scalacOptions ++= Seq(
-    // generate errors on unused imports
-    "-Xfatal-warnings",
-    "-Ywarn-unused",
-    // generate errors on procedure syntax
-    "-Wconf:cat=deprecation:e",
-    "-Xfuture",
-    "-deprecation"
-  ),
-  TaskKey[Unit]("checkLastCompilationCached") := {
-    val str = streams.value
-    val thisProject = thisProjectRef.value
-
-    // readText expects a fully resolved scope (no This)
-    val compileIncScope = Global.in(thisProject, Compile, compileIncremental.key)
-    val reader = str.readText(streamScopedKey(compileIncScope), None)
-    val lines = IO.readLines(reader)
-
-    val cachedCompilationLogEntry = "No changes"
-    if (lines.forall(line => !line.contains(cachedCompilationLogEntry))) {
-      lines.foreach(line => str.log.error(line))
-      throw new RuntimeException("last compilation was not fully cached")
-    }
-  }
+  scalaVersion := V.scala212
 )
 
 // 2.13.x is supported
@@ -68,7 +44,3 @@ TaskKey[Unit]("check") := {
   assert(scalacOptions.in(scala212).value.count(_ == "-Yrangepos") == 1)
   assert(scalacOptions.in(scala213).value.count(_ == "-Yrangepos") == 1)
 }
-
-// https://github.com/sbt/sbt/commit/dbb47b3ce822ff7ec25881dadd71a3b29e202273
-// must be outside a macro to workaround "Illegal dynamic reference: Def"
-def streamScopedKey(scope: Scope) = Def.ScopedKey(scope, Keys.streams.key)
